--- a/src/smsg/db.cpp
+++ b/src/smsg/db.cpp
@@ -23,7 +23,9 @@ prefixes
     pm      - purged message token
 */
 
-CCriticalSection cs_smsgDB;
+// Modern C++ Migration: Database operations are typically non-recursive
+// TODO: Consider verge::sync::SharedMutex for read-heavy database operations
+VergeStdMutex cs_smsgDB;
 leveldb::DB *smsgDB = nullptr;
 
 bool SecMsgDB::Open(const char *pszMode)