… DEVELOPMENT.md
… MODERNIZATION_SUMMARY.md
… PERFORMANCE_ANALYSIS.md
… PHASE3_PULL_REQUEST.md
… README.md
… src/smessage.cpp
… src/smessage.h
… src/smsg/rpcsmessage.cpp
… src/sync.h
… src/timedata.cpp
… src/torcontroller.h
… src/wallet/rpcdump.cpp
… src/warnings.cpp
… src/warnings.h