@justinvforvendetta 2025-08-19
that is xvgbase, the xvg branded erc20 token on base network