@yflop 2025-08-19
Thanks very much.
I'm glad to have one of the new tokens now! 🙂

On Wed, Aug 20, 2025, 12:47 AM sunerok ***@***.***> wrote:

> *justinvforvendetta* left a comment (vergecurrency/verge#1160)
> <https://github.com/vergecurrency/verge/pull/1160#issuecomment-3201478237>
>
> that is xvgbase, the xvg branded erc20 token on base network
>
> —
> Reply to this email directly, view it on GitHub
> <https://github.com/vergecurrency/verge/pull/1160#issuecomment-3201478237>,
> or unsubscribe
> <https://github.com/notifications/unsubscribe-auth/APZAWXKQO3VHV6OKDU2T4A33ONIJRAVCNFSM6AAAAAB77VREKSVHI2DSMVQWIX3LMV43OSLTON2WKQ3PNVWWK3TUHMZTEMBRGQ3TQMRTG4>
> .
> You are receiving this because you were mentioned.Message ID:
> ***@***.***>
>