@yflop 2025-08-20
Yep :)
grantpeace is my discord

On Thu, Aug 21, 2025 at 3:22 AM sunerok ***@***.***> wrote:

> *justinvforvendetta* left a comment (vergecurrency/verge#1160)
> <https://github.com/vergecurrency/verge/pull/1160#issuecomment-3207756236>
>
> do you have an email or discord?
>
> —
> Reply to this email directly, view it on GitHub
> <https://github.com/vergecurrency/verge/pull/1160#issuecomment-3207756236>,
> or unsubscribe
> <https://github.com/notifications/unsubscribe-auth/APZAWXNHHIHOFRWKDDMEWDD3OTDIZAVCNFSM6AAAAAB77VREKSVHI2DSMVQWIX3LMV43OSLTON2WKQ3PNVWWK3TUHMZTEMBXG42TMMRTGY>
> .
> You are receiving this because you were mentioned.Message ID:
> ***@***.***>
>